ACTIN-68: Refactoring
