+++ algo/src/main/java/com/hartwig/serve/sources/ckb/ActionableTrial.java
@@ -8,6 +8,6 @@ import org.jetbrains.annotations.Nullable;
 
 @Value.Immutable
 @Value.Style(passAnnotations = { NotNull.class, Nullable.class })
-public abstract class ActionableTrial implements ActionableEvent {
+abstract class ActionableTrial implements ActionableEvent {
 
 }
